Rename api/http/helpers files to kebab-case for consistency
