+++ api/http/helpers/delete-account.ts
