+++ api/test/rate-limiter.test.ts
@@ -2,9 +2,9 @@ import { describe, it, expect, beforeEach, vi } from 'vitest';
 
 // Mock the test-bypass check so bypass behavior is deterministic regardless of
 // the ambient TEST_BYPASS_SECRET / NODE_ENV. Defaults to "not bypassed".
-vi.mock('../http/helpers/checkTestBypass', () => ({ default: vi.fn(() => false) }));
+vi.mock('../http/helpers/check-test-bypass', () => ({ default: vi.fn(() => false) }));
 
-import checkTestBypass from '../http/helpers/checkTestBypass';
+import checkTestBypass from '../http/helpers/check-test-bypass';
 import { MemoryRateLimitStore } from '../http/rate-limit/store';
 import { RateLimiter, type RateLimitRequest } from '../http/rate-limit/rate-limiter';
 import { TooManyRequestsError } from '../http/errors/http-errors';
